# Implement hoare partition