Hello over there,

Thank you for this useful piece of code.

However, on my type of data I have found some kind of a problem which prevented the code to work as fast as it should be. In order to mitigate the problem, I implemented Hoare's partitioning and a new benchmark that tries to show the problem.

With my changes:

```
% go test -bench . -benchmem
goos: linux
goarch: amd64
pkg: github.com/Derlafff/psort
BenchmarkSort-4                      300           3524221 ns/op              64 B/op          2 allocs/op
BenchmarkPSort-4                    3000            412915 ns/op              64 B/op          2 allocs/op
BenchmarkPSortLowDisc-4             1000           1387552 ns/op              64 B/op          2 allocs/op
PASS
ok      github.com/Derlafff/psort       8.362s
```

Without my changes:

```
 % go test -bench . -benchmem                         
goos: linux
goarch: amd64
pkg: github.com/bluele/psort
BenchmarkSort-4                      500           3594622 ns/op              64 B/op          2 allocs/op
BenchmarkPSort-4                    3000            467009 ns/op              64 B/op          2 allocs/op
BenchmarkPSortLowDisc-4              300           5392569 ns/op              64 B/op          2 allocs/op
PASS
ok      github.com/bluele/psort 9.707s
```

I feel like there is a room for improvement both in the code and in the benchmark (it actually works much faster than on real data), but right now have no idea how to fix that.

Thank you.